Update ScoreReport's player ID keys

+++ pkg/game/test/factories.go
@@ -41,7 +41,7 @@ func GetTestPlacedTileWithMeeple(meeple elements.MeeplePlacement) elements.Place
 
 func GetTestScoreReport() elements.ScoreReport {
 	return elements.ScoreReport{
-		ReceivedPoints:  map[int]uint32{0: 5},
-		ReturnedMeeples: map[int][]uint8{},
+		ReceivedPoints:  map[uint8]uint32{0: 5},
+		ReturnedMeeples: map[uint8][]uint8{},
 	}
 }
